mlops-project/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
scikit-learn==1.7.0
pandas==2.2.3
numpy==2.2.6
mlflow==2.20.3
evidently==0.4.36
python-dotenv==1.0.1
pytest==8.3.5
httpx==0.28.1
streamlit==1.45.1
